ADDED: vector classes ADDED: waypoint and acceleration parsing to parser class

+++ src/data/Waypoint.cpp
@@ -1,6 +1,6 @@
 #include "Waypoint.h"
 
-Waypoint::Waypoint(const unsigned int time, const unsigned int longitude, const unsigned int latitude, const unsigned int altitude, const unsigned int speed)
+Waypoint::Waypoint(const unsigned int time, const double longitude, const double latitude, const double altitude, const double speed)
 	: m_time(time)
 	, m_longitude(longitude)
 	, m_latitude(latitude)
