+++ src/data/Waypoint.h
@@ -6,17 +6,17 @@
 class Waypoint
 {
 public:
-	Waypoint(const unsigned int time, const unsigned int longitude, const unsigned int latitude, const unsigned int altitude, const unsigned int speed);
+	Waypoint(const unsigned int time, const double longitude, const double latitude, const double altitude, const double speed);
 	~Waypoint();
 
 private:
 	unsigned int m_time;
 
-	unsigned int m_longitude;
-	unsigned int m_latitude;
+	double m_longitude;
+	double m_latitude;
 
-	unsigned int m_altitude;
-	unsigned int m_speed;
+	double m_altitude;
+	double m_speed;
 };
 
 #endif // WAYPOINT_H
